+++ crates/full-node/sov-sequencer/tests/integration/main.rs
@@ -8,6 +8,7 @@ mod rate_limits;
 mod sequencing_metadata;
 mod setup_mode;
 mod standard_sequencer;
+mod state_map_iteration;
 mod thin_sequencer;
 mod uniqueness;
 mod upgradability;